+++ src/tokenAuthenticator.ts
@@ -1,0 +1,7 @@
+import { Authenticate } from './types'
+
+export function tokenAuthenticator(token: string): Authenticate {
+  return () => ({
+    Authorization: `Bearer ${token}`,
+  })
+}