+++ src/types.ts
@@ -1,6 +1,6 @@
 import http from 'http'
 
-export type Authenticate = () => Promise<http.OutgoingHttpHeaders>
+export type Authenticate = () => http.OutgoingHttpHeaders
 
 export type OneReportResponseBody = {
   testSetExecutionId: string